add useRequiredAuth hook
